requirements.txt
alembic==1.14.0
asgiref==3.8.1
flask==3.1.0
flask-sqlalchemy==3.1.1
openai==1.58.1
psycopg2-binary==2.9.10
pydantic==2.10.4
pytest-flask==1.3.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
uvicorn==0.34.0
